+++ tests/test_integration.py
@@ -2,6 +2,7 @@ import sys
 import unittest
 import time
 from flask import Flask
+from queue import Queue
 
 # remove mocks and import flask_mqtt
 try:
@@ -9,7 +10,11 @@ try:
     sys.modules.pop('flask_mqtt')
 except KeyError:
     pass
-from flask_mqtt import Mqtt
+from flask_mqtt import Mqtt, MQTT_ERR_SUCCESS
+
+
+def wait(seconds=2):
+    time.sleep(seconds)
 
 
 class FlaskMQTTTestCase(unittest.TestCase):
@@ -30,17 +35,45 @@ class FlaskMQTTTestCase(unittest.TestCase):
 
     def test_subscribe(self):
         self.mqtt = Mqtt(self.app)
-        self.mqtt.subscribe('test')
-        self.assertEqual('test', self.mqtt.topics[0])
+        self.subscribe_handled = False
+        self.unsubscribe_handled = False
+
+        @self.mqtt.on_subscribe()
+        def handle_subscribe(client, userdata, mid_, granted_qos):
+            self.subscribe_handled = True
+
+        @self.mqtt.on_unsubscribe()
+        def handle_unsubscribe(client, userdata, mid_):
+            self.unsubscribe_handled = True
+
+        ret, mid = self.mqtt.subscribe('home/test')
+        self.assertEqual(ret, MQTT_ERR_SUCCESS)
+        wait()
+
+        ret, mid = self.mqtt.unsubscribe('home/test')
+        self.assertEqual(ret, MQTT_ERR_SUCCESS)
+        wait()
+
+        self.assertTrue(self.subscribe_handled)
+        self.assertTrue(self.unsubscribe_handled)
+
+    def test_topic_count(self):
+        self.mqtt = Mqtt(self.app)
+
+        ret, mid = self.mqtt.subscribe('test')
         self.assertEqual(1, len(self.mqtt.topics))
-        self.mqtt.subscribe('test')
+
+        ret, mid = self.mqtt.subscribe('test')
         self.assertEqual(1, len(self.mqtt.topics))
+
         self.mqtt.unsubscribe('test')
         self.assertEqual(0, len(self.mqtt.topics))
-        self.mqtt.unsubscribe('test')
 
-        self.mqtt.subscribe('test1')
-        self.mqtt.subscribe('test2')
+        self.mqtt.unsubscribe('test')
+        self.assertEqual(0, len(self.mqtt.topics))
+
+        ret, mid = self.mqtt.subscribe('test1')
+        ret, mid = self.mqtt.subscribe('test2')
         self.assertEqual(2, len(self.mqtt.topics))
         self.mqtt.unsubscribe_all()
         self.assertEqual(0, len(self.mqtt.topics))
@@ -49,18 +82,49 @@ class FlaskMQTTTestCase(unittest.TestCase):
     def test_publish(self):
 
         self.mqtt = Mqtt(self.app)
+        self.handled_message = False
+        self.handled_topic = False
+        self.handled_publish = False
 
         @self.mqtt.on_message()
         def handle_message(client, userdata, message):
-            print(message)
+            self.handled_message = True
 
-        @self.mqtt.on_topic('test')
+        @self.mqtt.on_publish()
+        def handle_publish(client, userdata, mid):
+            self.handled_publish = True
+
+        self.mqtt.subscribe('home/test')
+        wait()
+        self.mqtt.publish('home/test', 'hello world')
+        wait()
+
+        self.assertTrue(self.handled_message)
+        # self.assertTrue(self.handled_topic)
+        self.assertTrue(self.handled_publish)
+
+    def test_on_topic(self):
+
+        self.mqtt = Mqtt(self.app)
+        self.handled_message = False
+        self.handled_topic = False
+
+        @self.mqtt.on_message()
+        def handle_message(client, userdata, message):
+            self.handled_message = True
+
+        @self.mqtt.on_topic('home/test')
         def handle_on_topic(*args, **kwargs):
-            pass
+            self.handled_topic = True
 
-        self.mqtt.subscribe('test')
-        self.mqtt.publish('test', 'hello world')
-        time.sleep(2)
+        self.mqtt.subscribe('home/test')
+        wait()
+        self.mqtt.publish('home/test', 'hello world')
+        wait()
+
+        self.assertFalse(self.handled_message)
+        self.assertTrue(self.handled_topic)
+
 
     def test_logging(self):
         self.mqtt = Mqtt(self.app)
@@ -73,4 +137,5 @@ class FlaskMQTTTestCase(unittest.TestCase):
 
         self.mqtt.publish('test', 'hello world')
 
-
+if __name__ == '__main__':
+    unittest.main()